回复列表的展示

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -25,7 +25,7 @@ public class PublishController {
 
 
     @GetMapping("publish/{id}")
-    public  String edit(@PathVariable(name = "id")Integer id,
+    public  String edit(@PathVariable(name = "id")Long id,
                         Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
