+++ src/main/java/life/majiang/community/mapper/UserMapper.java
@@ -15,9 +15,11 @@ public interface UserMapper {
     @Select("select * from user where token = #{token}")
     User findByToken(@Param("token") String token);
     @Select("select * from user where id = #{id}")
-    User findById(@Param("id")Integer creator);
+    User findById(@Param("id")Long creator);
     @Select("select * from user where account_id = #{accountId}")
     User findByAccountId(@Param("accountId")String accountId);
     @Update("update user set name = #{name},token = #{token},gmt_modified = #{gmtModified},avatar_url=#{avatarUrl} where id = #{id}")
     void update(User user);
+
+
 }
